Fix Mattermost Webhook unable to edit messages.
Added logic for sendWebhook() to get the msgId of the post it created.
Required adding a custom field to the message's "props"-object.

Added logic to make mattermost's Send() method skip sendWebhook() if
 it received an edit-event.
Instead the regular Token-based EditMessage method will be called.

Editing with EditMessage is still broken in terms of user-spoofing,
 where if matterbridge edits a spoofed message,
 said message looses its spoofing entirely.
But this way users don't have to choose either spoofing or editing,
 and can instead have it spoof whenever possible,
 and simply fall back to no spoofing when it's not.